[Codechange] Move dynamic collision handling into separate file
