+++ source/main/CMakeLists.txt
@@ -220,6 +220,7 @@ set( SOURCE_FILES
   physics/air/TurboProp.{h,cpp}
   physics/collision/CartesianToTriangleTransform.h
   physics/collision/Collisions.{h,cpp}
+  physics/collision/DynamicCollisions.{h,cpp}
   physics/collision/PointColDetector.{h,cpp}
   physics/collision/Triangle.h
   physics/flex/Flexable.{h,cpp}
