+++ source/main/physics/collision/DynamicCollisions.h
@@ -1,0 +1,42 @@
+/*
+   This source file is part of Rigs of Rods
+   Copyright 2005-2012 Pierre-Michel Ricordel
+   Copyright 2007-2012 Thomas Fischer
+   Copyright 2016 Fabian Killus
+
+   For more information, see http://www.rigsofrods.com/
+
+   Rigs of Rods is free software: you can redistribute it and/or modify
+   it under the terms of the GNU General Public License version 3, as
+   published by the Free Software Foundation.
+
+   Rigs of Rods is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with Rigs of Rods.  If not, see <http://www.gnu.org/licenses/>.
+   */
+
+#pragma once
+
+class PointColDetector;
+class collcab_rate_t;
+class node_t;
+class Beam;
+class ground_model_t;
+
+void interTruckCollisions(const float dt, PointColDetector &interPointCD,
+        const int free_collcab, int collcabs[], int cabs[],
+        collcab_rate_t inter_collcabrate[], node_t nodes[],
+        const float collrange, Beam **trucks,
+        const int numtrucks,
+        ground_model_t &submesh_ground_model);
+
+void intraTruckCollisions(const float dt, PointColDetector &intraPointCD,
+        const int free_collcab, int collcabs[], int cabs[],
+        collcab_rate_t intra_collcabrate[], node_t nodes[],
+        const float collrange,
+        ground_model_t &submesh_ground_model);
+
